+++ R/.onAttach.R
@@ -1,0 +1,5 @@
+#' @title .onAttach
+#' @noRd
+.onAttach <- function(libname, pkgname) {
+  test_install()
+}