+++ R/getFluo.R
@@ -11,16 +11,32 @@
 #' @param cubeSize half the length of the cube side length. By default the cube is 10x10x10 pixels.
 #' @param H5_path path to the H5 file used to create the mamut.xml file
 #'
-#' @importFrom rhdf5 h5closeAll h5read
 #' @importFrom dplyr enquo filter select
 #' @importFrom glue glue
+#' @importFrom utils packageVersion
 #'
 #' @return a list of length equal to the number of timepoints. For each timepoint, values of fluorescence around each spot location
 #' @export
 #' @note Many thanks to Dr. Mike Smith for his help on this function.
 #'
 
-getFluo <- function(H5_path, Spots_df, x_px = x_px, y_px = y_px, z_px = z_px, timepoint = FRAME, cubeSize = 5) {
+getFluo <- function(H5_path, Spots_df, x_px = x_px, y_px = y_px, z_px = z_px,
+                    timepoint = FRAME, cubeSize = 5) {
+
+  if (!requireNamespace("rhdf5")) {
+    stop("get_fluo() requires {rhdf5}, here are the instruction to install it.",
+         "*Please note that it will only work with R > 3.5*\n",
+         "Please install {rhdf5} using:\n",
+         "install.packages(\"BiocManager\"); ",
+         "BiocManager::install(\"rhdf5\")")
+  } else if (packageVersion("rhdf5") < 2.24 ) {
+    stop("get_fluo() requires {rhdf5} > 2.24, here are the instruction to install it.\n",
+         "*Please note that it will only work with R > 3.5*\n",
+         "Please update {rhdf5} using:\n",
+         "install.packages(\"BiocManager\"); ",
+         "BiocManager::install(\"rhdf5\")")
+  }
+
   x_px <- enquo(x_px)
   y_px <- enquo(y_px)
   z_px <- enquo(z_px)
@@ -42,11 +58,11 @@ getFluo <- function(H5_path, Spots_df, x_px = x_px, y_px = y_px, z_px = z_px, ti
     cells <- lapply(S, function(coords, H5_path) {
       nameSpot <- coords$name
       coords <- as.integer(coords[2:4])
-      pixels <- h5read(file = H5_path, name = glue("/t00{t}/s00/0/cells"), #for small example, green channel in s01; check how are channels in bioutifoul
+      pixels <- rhdf5::h5read(file = H5_path, name = glue("/t00{t}/s00/0/cells"), #for small example, green channel in s01; check how are channels in bioutifoul
                        index = list((coords[1]-cubeSize):(coords[1]+cubeSize),
                                     (coords[2]-cubeSize):(coords[2]+cubeSize),
                                     (coords[3]-cubeSize):(coords[3]+cubeSize)))
-      h5closeAll()
+      rhdf5::h5closeAll()
       return(list(t, nameSpot, pixels))
     },
     H5_path)