+++ NEWS.md
@@ -1,6 +1,10 @@
 # mamut2r 0.0.2.9000
 
+* Add `test_install()` in .onAttach
+* Put {rhdf5} in Suggests and add an error in `getFluo()` if needed.
+
 # mamut2r 0.0.2
+
 * Cleaned examples
 * Zenodo release